Create challenge stats script
